The get_shared_info was incorrect, and now all work on xenbus is complete Signed-off-by: Daniel Castro <evil.dani@gmail.com>
The test_xenstore function lists the vbd-ids then creates an entry
in the device/vbd/ID/test then creates a watch on the new key
and waits for the event arrival, that event for now has to be create
by the user using xenstore-write. This is provided to prove that it
works as expected.